feat: re-style button actions & fix bugs
